dts: Select SoftDevice Controller DTS binding as default
Once nrf7120 is upstreamed this commit will be noup in sdk-zephyr.

The SoftDevice Controller is a different controller than
the open source link layer with a different set of quirks.
It should therefore have its own device tree binding.

This commit converts the SoftDevice Controller driver
to use this new DTS binding instead of reusing the existing one.

This commit updates or adds additional overlays for
existing samples, applications and tests that were using the
open source link layer.

Signed-off-by: Robert Robinson <robert.robinson@nordicsemi.no>