boards: nordic: Add non-secure targets for nRF7120
This adds the nrf7120pdk/cpuapp/ns board variant to sdk-zephyr.
It allows to build nrf7120pdk with TF-M in nRF Connect SDK.

Once upstreamed to zephyr, commit will become a noup within sdk-zephyr.

This is done as a noup as the out-of-tree board extension feature is
not available. When zephyrproject-rtos/zephyr#69548 is completed this
commit can be reverted and the variant be placed in sdk-nrf.

Signed-off-by: Robert Robinson <robert.robinson@nordicsemi.no>